http3: collect exported interfaces into interface.go
